ref(config-store): Track partitions of configs
Since we'll be using kafka to store and receive configurations, each
config will be part of a kafka partiton. These can be added or removed
from the checker during rebalancing, thus we need to support dropping a
set of configurations all at once when a partition is removed.

The ConfigStore has learned `drop_partition`, which drops all
CheckConfigs associated to a particular partition. It does this
efficiently by keeping a TickBuckets list for each partition. When a
partition is dropped the entire TickBuckets vector is dropped and all
CheckConfigs contained in those buckets are dropped