fix(uptime): Prevent configs for the same subscription from duplicating (#111)
If we have multiple values for a subscription in the config topic we
will end up with duplicate subscriptions. This pr makes sure that
subscriptions won't duplicate when we have an updated config in the
topic.

Previously, we decided that since we will always delete/create
subscriptions to update them that we didn't need to de-dupe. Due to the
way that we generate test configs for the uptime checker this caused a
lot of duplicates and confusion with testing. It seems safer to just
guarantee that we don't get duplicates.

As well as this, the way we send updates from sentry has an at least
once guarantee. So it's possible that we might send the same update more
than once in certain failure scenarios. This fix will also guard against
this.