fix: Run consuemr in it's own thread
We don't want this running as a tokio task since it will block and
consume an entire thread. Tokio will only have so many task executor threads so
we may end up the scenario where executors are available to drive
futures forward.

The reason it was run in a tokio task before was because the partitions
update function in the manager would attempt to start new tokio tasks,
which it could not do outside of the tokio runtime.

Instead we can use tokios `task::spawn_blocking` to run the consumer,
which tokio will run in it's own blocking thread.