feat(uptime): Store completed ticks in redis (#127)
This stores tick progress in Redis and loads it when the scheduler
starts up. We use this to set the instance that we start from, so that
we won't miss ticks during deploys, rebalances and crashes.

We store the ticks as nanos and use them to reset the Interval back in
time. We store the completed ticks in
`in_order_tick_processor_join_handle`, which gives us a guaranteed
in-order completion of ticks, so that we know we're not missing
anything.