feat(uptime): Make scheduler aware of regions, and able to skip checks when it's not a given region's turn.
Now that we're introducing multiple regions the scheduler needs to be aware of which region it's in, and of the regions in the configs. It can then only run checks for regions when it's its turn to do so.

To calculate whether it's time to run, we look at the current timestamp, and divide it by the interval size. This gets us the total number of time that interval has occurred. We then modulo that by the number of regions, to figure out which region's turn it is to run. Finally, with that number, we compare the current region to that and decide whether to run.