feat(uptime): Make sure result and trace ids are consistent for a scheduled check
We want to ensure that we produce the same ids for the same check. Each scheduled check is uniquely identified via the subscription id and scheduled time. Having these ids be consistent makes it easy for us to de-dupe data if we have duplicates due to a deploy or other issue.