feat: allow internal ips for self-hosted instances (#200)
For self-hosted use cases, we will be checking against private IP
address space, most of the times. Since most of the self-hosted
deployment are done in a private (and mostly air-gapped) environment.

This PR introduces `allow_internal_ips` on the configuration that
hopefully does not break with existing running instances.