fix: make max_idle_per_host 0 (#204)
we intermittently receive reports that we generate a downtime issue with
the decription being `IncompleteMessageError`, and our customers say
that their website was in fact up, not down. According to
https://github.com/hyperium/hyper/issues/2136, this could be due to
connection pooling / a race condition between the client and server. one
of the comments suggests setting the connection pool timeout to `0`.
https://github.com/hyperium/hyper/issues/2136#issuecomment-861826148

we do not need connection pooling as we aren't making that many requests
to one particular host, so there should be no effect, and we can monitor
if this fixes the error.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>