feat(uptime): Add redis config updater
This adds a method that will regularly poll redis for updates to the configs.

Note that I  use a dict here for the updates - since we're killing services and then bringing them back up, there's no need to co-ordinate here. Using a dict allows us to de-dupe updates so that only the last one will be processed for a given config.