ref(metrics): record num_running along with queue size (#211)
Prior to this queue_size was being recorded as `check_concurrency`,
which wasn't quite right, since we were recording the value when added
into the queue. Add a new metric for `num_running`, which is how many
things are being executed at once. This should never be larger than the
current concurrency limit