fix(uptime): Fix race bug for scheduler progress key
This fixes a bug that occurs when multiple checkers are competing for the same progress key. If a checker saves the key, and then another boots and gets a value too close to the current time then we end up hitting an `OutOfRangeError` when we subtract `start` from `Utc::Now()`.