fix(uptime): Fix slot distribution to use a `uuid5`
Our slot distribution is currently based on the subscription id. The problem with this is that `subscription_id` is already used to partition our configs. This has the side effect of causing all configs in a partition to be placed into the same schedling slots. The result here is that partitions run all at once, then completely stop for a few ticks, then run again, and son on. Using a uuid with a different namespace here will distribute them more evenly.