fix(uptime): Fix slot distribution to use a `uuid5` (#257)
Our slot distribution is currently based on the subscription id. The
problem with this is that `subscription_id` is already used to partition
our configs. This has the side effect of causing configs in a partition
to not be distributed evenly among scheduling slots. The result here is
that partitions run all at once, then completely stop for a few ticks,
then run again, and son on. Using a uuid with a different namespace here
will distribute them more evenly.

Some math examples:
Since we modulo subscription ids by 128, then for example, anything that
modulos to 51 goes in the same bucket.
For each of those ids in the 51 bucket, if we modulo those by 60 (the
number of seconds for an interval), they all map to these numbers
[3, 7, 11, 15, 19, 23, 27, 31, 35, 39, 43, 47, 51, 55, 59]
if i do it for items in partition 50 instead
[2, 6, 10, 14, 18, 22, 26, 30, 34, 38, 42, 46, 50, 54, 58]

Since these all end up in only 1/4 of the total buckets, we end up with
spikiness in our checks.