feat(metrics): Monitor executor task metrics (#271)
This should help us to understand if our http checker tasks are taking a
suspiciously long time to be scheduled or polled.

Our suspicion is that when we turn up execution concurrency we may be
overloading the tokio scheduler in a way such that we are very slow to
poll the http tasks. By the time we finally poll the task the reqwest
timeout may have been reached and the request will be considered timed
out.