fix(uptime): Have uptime-checker shut down gracefully on sigterm (#295)
The executor will block indefinitely until all feeding producers are
dropped. We weren't shutting down the partitioned services when a
shutdown was issued, but even with fixing that, the executor still holds
on to a producer that it uses for its own retries. So, just add a
`take_until` on the stream, awaiting a future that is triggered by the
cancel signal.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>