fix(uptime): Add timeouts for redis rpcs; process many tick results at once (#302)
Add a configurable (default 30s) connection/response timeouts, so that
we don't block indefinitely awaiting a response that never comes. For
the scheduler specifically, break out and log the individual connection
errors we might see, so that we just log timeouts, but retry connecting
if there's an actual connection drop.

Also in the scheduler, process many tick results at once using
recv_many, and then simply use the last one in the (ordered) vec of
results as the appropriate progress key.

Finally, update redis to 0.30.0.