ref(uptime): return Result<> structs and log/trace instead of panicking in the scheduler loop (#307)
Part 1: scheduler loop. I think this is the sane pattern we can repeat
in our code for our other tasks: no more panicking, return errors, and
the top-level function that is launched inside a tokio::spawn gets
access to the common `tasks_finished` channel that will receive the
return result from that function. That in turn goes to the receiver in
the app, which can then make a decision about what (if anything) to do
with the result. In this PR, if there's any kind of error, we'll just
panic and die, but in the future we could put in some selective retries.

This does not address all the looping tasks we have; we still have the
redis writer and the executor. Those can be dealt with in forthcoming
PRs. If a panic does occur in any of those tasks, the resulting channels
they use to communicate with are closed, and subsequent reads/writes to
them generate errors in the scheduler, which lead to the entire service
being brought down--it's just very circuitous and requires judicious log
tracing to understand why.