ref(uptime): clean up atomic ordering usage (#309)
We use Sequential Consistent ordering for a couple of atomic accesses in
the uptime checker. These are all for independent counters, so this kind
of requirement is overkill; relaxed works just fine, and is more
performant.

(See https://doc.rust-lang.org/nomicon/atomics.html for a human-readable
discussion around what atomic orderings can do/how they ought to be
used.)