ref(uptime): replace join task with join_all future (#314)
We create a task, per-scheduler loop, that simply awaits all the check
results. I don't think we need this: we can easily batch all the futures
we are awaiting into a single JoinAll future to be awaited in the redis
writer task, and since the redis writer task must process all results in
tick order, we aren't wasting time waiting for long futures--we were
going to wait for them in the other task, anyway.