feat(uptime): Implement new uptime-checker deploy pipeline (#414)
Code deploy counterpart of https://github.com/getsentry/ops/pull/17829

Uptime checker is a distributed system where work is partitioned across
all running pods. Each pod knows which checks it's responsible for based
on:
  1. Its pod number (0, 1, 2, ...)
  2. The total number of pods running

When we deploy a new version, pods restart one-by-one. During this
restart period, some checks might not run because the pod responsible
for them is restarting. This creates gaps in monitoring coverage.

To solve this problem, we're going to modify the deploy to use a canary
deployment for coverage.

We deploy to two separate StatefulSets:
  - Production: The main deployment that handles all monitoring checks
  - Canary: A second deployment that runs the new code in read-only mode

The deploy process for this looks like

  1. Deploy new code to canary
  2. Scale up canary.
3. Optionally, let the canary soak for a period of time, and validate
whether the new version is safe to deploy. This can be disabled, and in
this case, we'll deploy to production as soon as canary is running
  4. Deploy to production
  5. Scale canary back down

During the production rollout, canary pods are running alongside
production pods. Both sets of pods are responsible for the same checks.
This means when a production pod restarts, the corresponding canary pod
continues running those checks, which allows us to have no gaps in
coverage.

Canary runs in read-only mode, which means it only writes the current
watermark to Redis, but otherwise doesn't touch it. That allows us to
avoid cases where the different deploys stomp on each other.

One trade-off that we've made with this implementation is that in the
config, the canary replicas are set to 0. So when canary is scaled up
during deployment, the cluster state temporarily diverges from config.
This will show diffs if someone runs kubectl diff during deployment.
This will only apply during deploy, until the canary has been scaled
back down.