feat(uptime): add uptime assertions data model + basic compiler (#416)
There is both a data model and a runtime model that look very, very similar. They will likely diverge even more in the future. The data model is in the root of the module, the runtime model is under "compiled", with a basic compiler to go from data -> runtime model, and some tests testing the same.